+++ src/console/controllers/UsersController.php
@@ -8,12 +8,13 @@
 namespace craft\console\controllers;
 
 use Craft;
-use DateTime;
 use craft\console\Controller;
 use craft\elements\User;
 use craft\helpers\ArrayHelper;
 use craft\helpers\Console;
 use craft\helpers\UrlHelper;
+use DateTime;
+use yii\base\InvalidArgumentException;
 use yii\console\ExitCode;
 
 /**
@@ -24,12 +25,6 @@ use yii\console\ExitCode;
  */
 class UsersController extends Controller
 {
-    /**
-     * @var int|null The user's id.
-     * @since 3.8.0
-     */
-    public $id;
-
     /**
      * @var string|null The user’s email address.
      * @since 3.7.0
@@ -107,10 +102,6 @@ class UsersController extends Controller
             case 'set-password':
                 $options[] = 'password';
                 break;
-            case 'impersonate':
-                $options[] = 'id';
-                $options[] = 'email';
-                $options[] = 'username';
         }
 
         return $options;
@@ -243,15 +234,15 @@ class UsersController extends Controller
     /**
      * Deletes a user.
      *
-     * @param string $usernameOrEmail The user’s username or email address.
+     * @param string $user The ID, username, or email address of the user account.
      * @return int
      */
-    public function actionDelete(string $usernameOrEmail): int
+    public function actionDelete(string $user): int
     {
-        $user = Craft::$app->getUsers()->getUserByUsernameOrEmail($usernameOrEmail);
-
-        if (!$user) {
-            $this->stderr("No user exists with a username/email of “{$usernameOrEmail}”." . PHP_EOL, Console::FG_RED);
+        try {
+            $user = $this->_user($user);
+        } catch (InvalidArgumentException $e) {
+            $this->stderr($e->getMessage() . PHP_EOL, Console::FG_RED);
             return ExitCode::UNSPECIFIED_ERROR;
         }
 
@@ -274,14 +265,14 @@ class UsersController extends Controller
                 return ExitCode::UNSPECIFIED_ERROR;
             }
 
-            if (!$this->confirm("Delete user “{$usernameOrEmail}” and transfer their content to user “{$this->inheritor}”?")) {
+            if (!$this->confirm("Delete user “{$user->username}” and transfer their content to user “{$inheritor->username}”?")) {
                 $this->stdout('Aborting.' . PHP_EOL);
                 return ExitCode::USAGE;
             }
 
             $user->inheritorOnDelete = $inheritor;
         } else if ($this->interactive) {
-            $this->deleteContent = $this->confirm("Delete user “{$usernameOrEmail}” and their content?");
+            $this->deleteContent = $this->confirm("Delete user “{$user->username}” and their content?");
 
             if (!$this->deleteContent) {
                 $this->stdout('Aborting.' . PHP_EOL);
@@ -308,15 +299,15 @@ class UsersController extends Controller
     /**
      * Changes a user’s password.
      *
-     * @param string $usernameOrEmail The user’s username or email address
+     * @param string $user The ID, username, or email address of the user account.
      * @return int
      */
-    public function actionSetPassword(string $usernameOrEmail): int
+    public function actionSetPassword(string $user): int
     {
-        $user = Craft::$app->getUsers()->getUserByUsernameOrEmail($usernameOrEmail);
-
-        if (!$user) {
-            $this->stderr("No user exists with a username/email of “{$usernameOrEmail}”." . PHP_EOL, Console::FG_RED);
+        try {
+            $user = $this->_user($user);
+        } catch (InvalidArgumentException $e) {
+            $this->stderr($e->getMessage() . PHP_EOL, Console::FG_RED);
             return ExitCode::UNSPECIFIED_ERROR;
         }
 
@@ -344,39 +335,61 @@ class UsersController extends Controller
     /**
      * Generate a URL to impersonate a user.
      *
+     * @param string $user The ID, username, or email address of the user account.
      * @return int
      */
-    public function actionImpersonate(): int
+    public function actionImpersonate(string $user): int
     {
-        if ($this->id !== null) {
-            $user = Craft::$app->getUsers()->getUserById($this->id);
-        }
-        elseif ($this->username !== null || $this->email !== null) {
-            $user = Craft::$app->getUsers()->getUserByUsernameOrEmail($this->username ?? $this->email);
-        }
-        
-        if (!$user) {
-            $this->stderr('No user found.');
+        try {
+            $user = $this->_user($user);
+        } catch (InvalidArgumentException $e) {
+            $this->stderr($e->getMessage() . PHP_EOL, Console::FG_RED);
             return ExitCode::UNSPECIFIED_ERROR;
         }
 
         $token = Craft::$app->getTokens()->createToken([
-            'users/i personate-with-token', [
+            'users/impersonate-with-token', [
                 'userId' => $user->id,
                 'prevUserId' => $user->id,
             ],
-        ], 1, new DateTime('+10 minutes'));
+        ], 1, new DateTime('+1 hour'));
 
         if (!$token) {
-            $this->stderr('Unable to create the impersonation token');
+            $this->stderr('Unable to create the impersonation token.' . PHP_EOL, Console::FG_RED);
             return ExitCode::UNSPECIFIED_ERROR;
         }
 
         $url = $user->can('accessCp') ? UrlHelper::cpUrl() : UrlHelper::siteUrl();
         $url = UrlHelper::urlWithToken($url, $token);
 
-        $this->output($url);
+        $this->stdout("Impersonation URL for $user->username: ");
+        $this->stdout($url . PHP_EOL, Console::FG_CYAN);
+        $this->stdout('(Expires in one hour.)' . PHP_EOL, Console::FG_GREY);
 
         return ExitCode::OK;
+    }
+
+    /**
+     * Resolves a `user` argument.
+     *
+     * @param string $value The `user` argument value
+     * @return User
+     * @throws InvalidArgumentException if the user could not be found
+     */
+    private function _user(string $value): User
+    {
+        if (is_numeric($value)) {
+            $user = Craft::$app->getUsers()->getUserById($value);
+            if (!$user) {
+                throw new InvalidArgumentException("No user exists with the ID: $value");
+            }
+        } else {
+            $user = Craft::$app->getUsers()->getUserByUsernameOrEmail($value);
+            if (!$user) {
+                throw new InvalidArgumentException("No user exists with the username/email: $value");
+            }
+        }
+
+        return $user;
     }
 }